+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### `latest`
+### `0.8`
 
 - Feat: #13 pass old value and all the current values to `on_event`
 - Feat: #14 `send_animation` method for Telegram bots